requirements.txt
flask
postgres
python-dotenv
requests